# ssa: removes map use for block traversals